Started asset manager, decided on structure
