+++ Core/Cargo.toml
@@ -49,6 +49,7 @@ num-traits.workspace = true
 hashbrown.workspace = true
 bitflags.workspace = true
 parking_lot.workspace = true
+crossbeam-channel.workspace = true
 
 # specific dependencies
 include_dir.workspace = true
@@ -56,6 +57,7 @@ shaderc.workspace = true
 openal.workspace = true
 image.workspace = true
 glam.workspace = true
+gltf.workspace = true
 
 # rendering dependencies
 winit.workspace = true
