+++ Core/src/render/backend/vulkan/buffer.rs
@@ -113,7 +113,11 @@ impl VkBuffer {
                 self.map();
             }
 
+            unsafe { self.mapped = self.mapped.offset(offset as isize) };
+
             unsafe { std::ptr::copy_nonoverlapping(ptr, self.mapped, self.buffer_size as usize) };
+
+            unsafe { self.mapped = self.mapped.offset(-(offset as isize)) };
 
             if need_to_map_buffer {
                 self.unmap()
