+++ Core/src/render/model.rs
@@ -1,0 +1,10 @@
+pub struct Model {
+
+}
+
+pub struct Material {
+    color: glam::Vec4,
+    emissive: glam::Vec4,
+    metallic: f32,
+    roughness: f32,
+}
