+++ Core/src/render/texture.rs
@@ -1,0 +1,3 @@
+pub struct Texture {
+
+}
